requirements.txt
streamlit
pandas
matplotlib.pyplot
seaborn
matplotlib.font_manager
os